refactor(analytics): update code to reflect package relocation of `MetricsContext`
